Count tokens for prompt to provide to modelRequirements.
This allows us in the default to ahve a smaller model, and then use a bigger one if necessary.

NOTE: this currently only works if the model htat is selected is of the same provider. Once a new provider
(e.g. anthropic) is mixed in it might no longer work.