Fix a very frustrating bug about streaming user messages.
This has been absurdly hard to diagnose!

The problem is when streaming parsing and getting a `\` followed by a `n`. The first part was escaped to \",
but then when the next one came in and knew it was a \n, it was now different than what had come in before.

We've added a test that fails before the change and succeeds after it to help lock in this behavior.

Part of #14.