Make it so there's only one _compiledData request a ta time for a sprout.
There is a problem where when the currentSprout changes, multipel views ask for multipel properties that all
rely on knowing the compiled result immediately. They would all kick off at the same time, kicking off many
networking requests, duplicate LLM calls, etc.

The solution is to have a layer of indirection that keeps track of the fetching once and returns that.

This massively reduces the extra fetch problem, but really should be applied throughtout any promise-ified
sprout methods.

Part of #15.