Fix some errors in the browser.fetch implementation.
The previous one wasn't exercised before because of the single fetch/compile response.

The previous one had a problem where the response it woudl vend was also the one that needed to be cloned,
but by the time it was vended a second time it might have already been used.

Now we store a promise signal to check the response.

Part of #15. Part of #7.