Tweak a weird race issue in sprout-view.
Before, sproutChanged only took the lastSprout, and whatever the currentSprout was now.

But this led to weird issues if _currentSprout changed multiple times before the next timer tick... as happens on first run when multiple API Keys are set (successive calls to setAPIKey). This could lead to sproutChanged with a last Sprout that was never actually seen in sproutChanged.

Now we snapshot both the current sprout and the last sprout, which is at least a littl emore sane.

Realistically we should probably batch up the setAPIKeys updates into one as well.

Part of #5.