Fetchers get writeSprout and deleteSprout.
They should fail if writeable(sproutName) is false.

This is a kind of coneceptual undo of #20.

DataManager in app now gets much more simple.

Now that we have MetaFetcher, it's cleaner for the same fetcher that knows how to fetch a given subdirectory knows how to write to those, too.

Fixes #22.